+++ source/databricks/package/infrastructure.py
@@ -14,8 +14,8 @@
 
 # Resource names and variables defined in the infrastructure repository (https://github.com/Energinet-DataHub/dh3-infrastructure)
 
-from package.codelists.time_series_type import TimeSeriesType
-from package.codelists.market_role import MarketRole
+from package.codelists import TimeSeriesType
+from typing import Union
 
 WHOLESALE_CONTAINER_NAME = "wholesale"
 
@@ -36,21 +36,23 @@ def get_container_root_path(storage_account_name: str) -> str:
 def get_result_file_relative_path(
     batch_id: str,
     grid_area: str,
-    gln: str,
+    gln: Union[str, None],
     time_series_type: TimeSeriesType,
+    grouping: str,
 ) -> str:
     batch_path = get_batch_relative_path(batch_id)
-    return f"{batch_path}/{RESULT_FOLDER}/grid_area={grid_area}/gln={gln}/time_series_type={time_series_type.value}"
+    relative_path = f"{batch_path}/{RESULT_FOLDER}/grouping={grouping}/time_series_type={time_series_type.value}/grid_area={grid_area}"
+    if gln is None:
+        return relative_path
+    else:
+        return f"{relative_path}/gln={gln}"
 
 
 def get_actors_file_relative_path(
-    batch_id: str,
-    grid_area: str,
-    time_series_type: TimeSeriesType,
-    market_role: MarketRole,
+    batch_id: str, grid_area: str, time_series_type: TimeSeriesType
 ) -> str:
     batch_path = get_batch_relative_path(batch_id)
-    return f"{batch_path}/{ACTORS_FOLDER}/grid_area={grid_area}/time_series_type={time_series_type.value}/market_role={market_role.value}"
+    return f"{batch_path}/{ACTORS_FOLDER}/time_series_type={time_series_type.value}/grid_area={grid_area}"
 
 
 def get_time_series_quarter_relative_path(