+++ source/databricks/tests/datamigration/migration_scripts/test_202212051100_Move_Results_To_Wholesale_Container.py
@@ -25,10 +25,9 @@ from package.datamigration.migration_script_args import MigrationScriptArgs
 def test__apply__directory_client_contructed_with_correct_arguments(
     mock_directory_client,
 ):
-
     # Arrange
     sut = get_migration_script()
-    migration_args = MigrationScriptArgs("", "", None)
+    migration_args = MigrationScriptArgs("", "", "", None)
     processes_container = "processes"
     results_source_directory = "results"
 
@@ -47,10 +46,9 @@ def test__apply__directory_client_contructed_with_correct_arguments(
 def test__apply__calls_rename_directory_with_correct_arguments(
     mock_directory_client,
 ):
-
     # Arrange
     sut = get_migration_script()
-    migration_args = MigrationScriptArgs("", "", None)
+    migration_args = MigrationScriptArgs("", "", "", None)
     mock_directory_client.return_value.exists.return_value = True
 
     # Act
@@ -68,10 +66,9 @@ def test__apply__calls_rename_directory_with_correct_arguments(
 def test__apply__when_source_directory_not_exist__never_call_rename_directory(
     mock_directory_client,
 ):
-
     # Arrange
     sut = get_migration_script()
-    migration_args = MigrationScriptArgs("", "", None)
+    migration_args = MigrationScriptArgs("", "", "", None)
     mock_directory_client.return_value.exists.return_value = False
 
     # Act