+++ source/databricks/tests/datamigration/migration_scripts/test_202301021400_Rename_Results_Folder_To_Calculation-output.py
@@ -25,10 +25,9 @@ from package.datamigration.migration_script_args import MigrationScriptArgs
 def test__apply__directory_client_contructed_with_correct_arguments(
     mock_directory_client: Mock,
 ) -> None:
-
     # Arrange
     sut = get_migration_script()
-    migration_args = MigrationScriptArgs("", "", ANY)
+    migration_args = MigrationScriptArgs("", "", "", ANY)
     source_container = "wholesale"
     source_directory = "results"
 
@@ -47,10 +46,9 @@ def test__apply__directory_client_contructed_with_correct_arguments(
 def test__apply__calls_rename_directory_with_correct_arguments(
     mock_directory_client: Mock,
 ) -> None:
-
     # Arrange
     sut = get_migration_script()
-    migration_args = MigrationScriptArgs("", "", ANY)
+    migration_args = MigrationScriptArgs("", "", "", ANY)
 
     # Act
     sut.apply(migration_args)
@@ -67,10 +65,9 @@ def test__apply__calls_rename_directory_with_correct_arguments(
 def test__apply__when_source_directory_not_exist__never_call_rename_directory(
     mock_directory_client: Mock,
 ) -> None:
-
     # Arrange
     sut = get_migration_script()
-    migration_args = MigrationScriptArgs("", "", ANY)
+    migration_args = MigrationScriptArgs("", "", "", ANY)
     mock_directory_client.return_value.exists.return_value = False
 
     # Act